# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LensFlow AI - Real estate video pipeline API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings/avatar:
    get:
      tags: [settings]
      operationId: getAvatarSettings
      summary: Get the current user's saved digital-twin avatar
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Avatar settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvatarSettings'
        '401':
          description: Not authenticated
    put:
      tags: [settings]
      operationId: updateAvatarSettings
      summary: Save the current user's digital-twin avatar
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvatarSettings'
      responses:
        '200':
          description: Avatar settings saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvatarSettings'
        '401':
          description: Not authenticated
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — Bearer <sid>.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    AvatarSettings:
      type: object
      properties:
        heygenAvatarId:
          type: string
          description: HeyGen custom avatar ID (digital twin)
        heygenAvatarName:
          type: string
          description: Display name for this avatar (e.g. "Sarah's Twin")
        heygenVoiceId:
          type: string
          description: HeyGen voice ID to use with this avatar